trip-planner/backend/pyproject.toml
[project]
name = "ai-agent-backend"
version = "0.1.0"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]",
  "langchain-core",
  "langchain-openai",
  "pydantic>=2.0",
  "pydantic-settings",
  "python-dotenv",
]

[tool.uv]
dev-dependencies = [
  "pytest",
]
